pyproject.toml
[project]
name = "telegram-sorter"
version = "0.1.0"
description = "AI-powered Telegram news channel sorter"
requires-python = ">=3.9"
dependencies = [
    "telethon>=1.36.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "redis[asyncio]>=5.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

